@@ -19,7 +19,7 @@ dependencies:
   #  日志打印
   logger: '>=1.0.0 <2.0.0'
   #  网络请求相关dio
-  dio: '>=4.0.0 <5.0.0'
+  dio: '5.1.1'
   #  下拉刷新
   pull_to_refresh: '>=2.0.0 <3.0.0'
   synchronized: '>=3.0.0 <4.0.0'
@@ -28,16 +28,16 @@ dependencies:
   #  toast工具类
   bot_toast: '>=4.0.1'
   #  统一工具类
-  common_utils: '>=2.0.2'
-  decimal: '<2.0.0'
+  common_utils: '>=2.1.0'
+  decimal: '2.3.2'
   #  自适应/响应式--判定屏幕尺寸
-  adaptive_breakpoints: '>=0.0.5'
+  adaptive_breakpoints: '0.1.6'
   #  组件生命周期
-  lifecycle: '>=0.4.1 <0.5.0'
+  lifecycle: '0.6.0'
   # WebView
-  webview_flutter: '>=2.3.1'
+  webview_flutter: '3.0.4'
   #  设备信息
-  device_info_plus: '>=3.1.1'
+  device_info_plus: '>=3.1.1 <6.0.0'
   #  应用包信息
   package_info_plus: '>=1.3.0'
   # 文件路径获取
@@ -45,7 +45,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
